cambios varios modulo estados de pagos

+++ resources/views/livewire/finance/sigfe-archivo-compromiso.blade.php
@@ -1,13 +1,14 @@
-<div>
-    {{-- If you look to others for fulfillment, you will never truly be fulfilled. --}}
+<div>    
     @if ($archivoCompromiso !== null)
         <div class="mb-2">
             <button wire:click="downloadFile('{{ $archivoCompromiso }}')" class="btn btn-sm btn-success">
                 <i class="fas fa-download"></i> Descargar
             </button>
-            <button wire:click="deleteFile('{{ $archivoCompromiso }}')" class="btn btn-sm btn-danger">
-                <i class="fas fa-trash-alt"></i> Eliminar
-            </button>
+            @if(!$onlyRead)
+                <button wire:click="deleteFile('{{ $archivoCompromiso }}')" class="btn btn-sm btn-danger">
+                    <i class="fas fa-trash-alt"></i> Eliminar
+                </button>
+            @endif
         </div>
     @else
         <input type="file" wire:model="file" accept=".pdf">
